refactor(repo): update tooling and cleaning up builds
