+++ src/viewable.svelte
@@ -61,10 +61,10 @@
    * @type {String}
    */
   export let rootMargin = '0px';
-   /**
-    * Array of visibility thresholds that will result in a callback when
-    * the observed element crosses that each threshold (.1 = 10% visible inside of its container)
-    */
+  /**
+   * Array of visibility thresholds that will result in a callback when
+   * the observed element crosses that each threshold (.1 = 10% visible inside of its container)
+   */
   export let threshold = [0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1];
   /**
    * Observed element metadata
